pyproject.toml
[project]
name = "storytime"
version = "0.3.0"
description = "AI-powered children's book illustration generator using Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "streamlit>=1.40.0",
    # AI and Google GenAI
    "google-genai>=0.8.0",
    # Data models and validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Image processing[]
    "pillow>=10.0.0",
    # PDF generation
    "reportlab>=4.0.0",
    # HTTP requests
    "requests>=2.31.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.6",
    "structlog>=25.4.0",
    # LangChain and AI agents
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langchain-community>=0.3.27",
    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
